Deprecated Build Functions

renderthis, under the name xaringanBuilder, previously provided the same functionality using build_ functions. To be consistent with the new package name, these function names have also been changed.

build_html() is now to_html()

build_pdf() is now to_pdf()

build_png() is now to_png()

build_gif() is now to_gif()

build_mp4() is now to_mp4()

build_pptx() is now to_pptx()

build_social() is now to_social()

Argument names. Note that the input and output_file arguments of these functions have also been renamed. They are now named from and to.

Usage

build_html(...)

build_pdf(...)

build_png(...)

build_gif(...)

build_mp4(...)

build_pptx(...)

build_social(...)

Arguments

...

Parameters passed to the new to_*() function

Value

See the corresponding new function for the appropriate return value.